Replace switch statement with if
We only had two cases, and only one action.  It's much simpler as an "if".

+++ src/main/org/tvrenamer/view/ResultsTable.java
@@ -858,20 +858,11 @@ public final class ResultsTable implements Observer, AddEpisodeListener {
             @Override
             public void keyReleased(KeyEvent e) {
                 super.keyReleased(e);
-
-                switch (e.keyCode) {
-
-                    // backspace
-                    case '\u0008':
-                    // delete
-                    case '\u007F':
-                        deleteSelectedTableItems();
-                        break;
-
-                    // Code analysis says have a default clause...
-                    default:
+                if ((e.keyCode == '\u0008') // backspace
+                    || (e.keyCode == '\u007F')) // delete
+                {
+                    deleteSelectedTableItems();
                 }
-
             }
         });
